Move web-incompatible collection:collection tests to a dedicated test source set
