+++ collection/collection/src/commonTest/kotlin/androidx/collection/FloatFloatMapTest.kt
@@ -16,6 +16,7 @@
 
 package androidx.collection
 
+import kotlin.js.JsName
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
@@ -585,21 +586,6 @@ internal class FloatFloatMapTest {
     }
 
     @Test
-    fun equals() {
-        val map = MutableFloatFloatMap()
-        map[1f] = 1f
-
-        assertFalse(map.equals(null))
-        assertEquals(map, map)
-
-        val map2 = MutableFloatFloatMap()
-        assertNotEquals(map, map2)
-
-        map2[1f] = 1f
-        assertEquals(map, map2)
-    }
-
-    @Test
     fun containsKey() {
         val map = MutableFloatFloatMap()
         map[1f] = 1f
@@ -615,6 +601,22 @@ internal class FloatFloatMapTest {
 
         assertTrue(1f in map)
         assertFalse(2f in map)
+    }
+
+    @Test
+    @JsName("testEquals")
+    fun equals() {
+        val map = MutableFloatFloatMap()
+        map[1f] = 1f
+
+        assertFalse(map.equals(null))
+        assertEquals(map, map)
+
+        val map2 = MutableFloatFloatMap()
+        assertNotEquals(map, map2)
+
+        map2[1f] = 1f
+        assertEquals(map, map2)
     }
 
     @Test
