+++ collection/collection/src/commonTest/kotlin/androidx/collection/FloatObjectMapTest.kt
@@ -16,6 +16,7 @@
 
 package androidx.collection
 
+import kotlin.js.JsName
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFalse
@@ -622,6 +623,7 @@ internal class FloatObjectMapTest {
     }
 
     @Test
+    @JsName("testEquals")
     fun equals() {
         val map = MutableFloatObjectMap<String?>()
         map[1f] = "World"
