+++ collection/collection/src/commonTest/kotlin/androidx/collection/IntSetTest.kt
@@ -15,6 +15,7 @@
  */
 package androidx.collection
 
+import kotlin.js.JsName
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
@@ -376,6 +377,7 @@ internal class IntSetTest {
     }
 
     @Test
+    @JsName("testEquals")
     fun equals() {
         val set = MutableIntSet()
         set += 1
