+++ collection/collection/src/commonTest/kotlin/androidx/collection/LongSetTest.kt
@@ -15,6 +15,7 @@
  */
 package androidx.collection
 
+import kotlin.js.JsName
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
@@ -376,6 +377,7 @@ internal class LongSetTest {
     }
 
     @Test
+    @JsName("testEquals")
     fun equals() {
         val set = MutableLongSet()
         set += 1L
