+++ collection/collection/src/commonTest/kotlin/androidx/collection/ObjectFloatMapTest.kt
@@ -16,6 +16,7 @@
 
 package androidx.collection
 
+import kotlin.js.JsName
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
@@ -601,6 +602,7 @@ internal class ObjectFloatTest {
     }
 
     @Test
+    @JsName("testEquals")
     fun equals() {
         val map = MutableObjectFloatMap<String?>()
         map["Hello"] = 1f
