+++ collection/collection/src/commonTest/kotlin/androidx/collection/ObjectLongMapTest.kt
@@ -16,6 +16,7 @@
 
 package androidx.collection
 
+import kotlin.js.JsName
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
@@ -601,6 +602,7 @@ internal class ObjectLongTest {
     }
 
     @Test
+    @JsName("testEquals")
     fun equals() {
         val map = MutableObjectLongMap<String?>()
         map["Hello"] = 1L
