+++ collection/collection/src/commonTest/kotlin/androidx/collection/ScatterMapTest.kt
@@ -16,6 +16,7 @@
 
 package androidx.collection
 
+import kotlin.js.JsName
 import kotlin.test.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertFailsWith
@@ -683,6 +684,7 @@ internal class ScatterMapTest {
     }
 
     @Test
+    @JsName("testEquals")
     fun equals() {
         val map = MutableScatterMap<String?, String?>()
         map["Hello"] = "World"
